// Project_Apex/frontend/src/api/models/UserNotificationPreferencesPublic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserNotificationPreferencesPublic = {
    email_notifications: boolean;
    browser_notifications: boolean;
    copy_trading_alerts: boolean;
    withdrawal_alerts: boolean;
    market_updates: boolean;
    security_alerts: boolean;
    user_id: string;
    updated_at: string;
};