// Project_Apex/frontend/src/api/models/UserNotificationPreferencesUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UserNotificationPreferencesUpdate = {
    email_notifications?: boolean | null;
    browser_notifications?: boolean | null;
    copy_trading_alerts?: boolean | null;
    withdrawal_alerts?: boolean | null;
    market_updates?: boolean | null;
    security_alerts?: boolean | null;
};